backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
ansible-runner==2.3.6
boto3==1.34.34
jinja2==3.1.3
pyyaml==6.0.1
kubernetes==29.0.0
sse-starlette==2.0.0
python-multipart==0.0.6
cryptography==42.0.0
